// frontend/src/app/api-services/api/model/task-status.ts
/**
 * StudyConnect Backend API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum TaskStatus {

    Open = 'OPEN',

    InProgress = 'IN_PROGRESS',

    Completed = 'COMPLETED',

    Cancelled = 'CANCELLED'
}